refactor(webfont-generator): split result domain
