+++ packages/webfont-generator/src/types.rs
@@ -1,19 +1,10 @@
-use std::collections::{BTreeMap, HashMap};
-use std::sync::Arc;
-use std::sync::{Mutex, OnceLock};
+use std::collections::HashMap;
 
-#[cfg(feature = "napi")]
-use napi::bindgen_prelude::Uint8Array;
 #[cfg(feature = "napi")]
 use napi_derive::napi;
 use serde_json::{Map, Value};
 
-use crate::input::LoadedSvgFile;
-use crate::pipeline::TtfGlyphCache;
-use crate::rendering::{CachedTemplateData, CarriedRenderCache};
-use crate::svg::types::GlyphCache;
-
-/// What happened to a file, for [`GenerateWebfontsResult::regenerate`]. `name` is the
+/// What happened to a file, for [`crate::GenerateWebfontsResult::regenerate`]. `name` is the
 /// caller-resolved glyph name (the `rename` callback, if any, is applied by the caller).
 pub enum GlyphChange {
     /// A new file. `name` overrides the file-stem glyph name when `Some`.
@@ -296,445 +287,4 @@ pub struct GenerateWebfontsOptions {
     /// Whether to write generated files to disk. Set to `false` for
     /// in-memory usage. Defaults to `true`.
     pub write_files: Option<bool>,
-}
-
-#[derive(Clone)]
-pub(crate) struct ResolvedGenerateWebfontsOptions {
-    pub ascent: Option<f64>,
-    pub center_horizontally: Option<bool>,
-    pub center_vertically: Option<bool>,
-    pub css: bool,
-    pub css_dest: String,
-    pub css_template: Option<String>,
-    /// Fully-resolved codepoints for the current glyph set (explicit + auto-assigned). Rebuilt
-    /// by `finalize_generate_webfonts_options` from `explicit_codepoints` whenever the set changes.
-    pub codepoints: BTreeMap<String, u32>,
-    /// The user-supplied codepoints, kept as the stable base so re-resolving after an
-    /// incremental add/remove assigns the same codepoints a fresh build would.
-    pub explicit_codepoints: BTreeMap<String, u32>,
-    pub css_fonts_url: Option<String>,
-    pub descent: Option<f64>,
-    pub dest: String,
-    pub files: Vec<String>,
-    pub fixed_width: Option<bool>,
-    pub format_options: Option<FormatOptions>,
-    pub html: bool,
-    pub incremental: bool,
-    pub html_dest: String,
-    pub html_template: Option<String>,
-    pub font_height: Option<f64>,
-    pub font_name: String,
-    pub font_style: Option<String>,
-    pub font_weight: Option<String>,
-    pub ligature: bool,
-    pub normalize: bool,
-    pub order: Vec<FontType>,
-    pub optimize_output: Option<bool>,
-    pub preserve_aspect_ratio: Option<bool>,
-    pub round: Option<f64>,
-    pub start_codepoint: u32,
-    pub template_options: Option<Map<String, Value>>,
-    pub types: Vec<FontType>,
-    pub write_files: bool,
-}
-
-/// Rendered bytes for each requested output format. Held by [`GenerateWebfontsResult`] and
-/// produced by the generator's format pipeline; grouping them lets an incremental regenerate
-/// refresh every format in a single assignment.
-#[derive(Clone, Default)]
-pub(crate) struct FontOutputs {
-    pub(crate) svg_font: Option<Arc<String>>,
-    pub(crate) ttf_font: Option<Arc<Vec<u8>>>,
-    pub(crate) woff_font: Option<Arc<Vec<u8>>>,
-    pub(crate) woff2_font: Option<Arc<Vec<u8>>>,
-    pub(crate) eot_font: Option<Arc<Vec<u8>>>,
-}
-
-pub(crate) struct RegenerationState {
-    pub(crate) caches_dirty: bool,
-    pub(crate) glyph_cache: GlyphCache,
-    pub(crate) ttf_cache: Option<TtfGlyphCache>,
-    pub(crate) written_outputs: HashMap<String, [u8; 16]>,
-}
-
-pub(crate) struct RegenerationStateLease {
-    slot: Arc<Mutex<Option<RegenerationState>>>,
-    state: Option<RegenerationState>,
-    keep_caches: bool,
-}
-
-/// Failure from asynchronous regeneration. Ordinary regeneration errors retain the input result
-/// so callers can recover it with [`RegenerateError::into_result`] and retry.
-pub struct RegenerateError {
-    result: Option<Box<GenerateWebfontsResult>>,
-    source: std::io::Error,
-}
-
-impl RegenerateError {
-    pub(crate) fn new(result: Option<GenerateWebfontsResult>, source: std::io::Error) -> Self {
-        Self {
-            result: result.map(Box::new),
-            source,
-        }
-    }
-
-    /// Return the result that was consumed by the failed operation. This is `None` only if the
-    /// blocking task was cancelled by the runtime before it could return the result.
-    pub fn into_result(self) -> Option<GenerateWebfontsResult> {
-        self.result.map(|result| *result)
-    }
-
-    /// Return both the recoverable result and the underlying I/O error.
-    pub fn into_parts(self) -> (Option<GenerateWebfontsResult>, std::io::Error) {
-        (self.result.map(|result| *result), self.source)
-    }
-}
-
-impl std::fmt::Debug for RegenerateError {
-    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter
-            .debug_struct("RegenerateError")
-            .field("recoverable", &self.result.is_some())
-            .field("source", &self.source)
-            .finish()
-    }
-}
-
-impl std::fmt::Display for RegenerateError {
-    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.source.fmt(formatter)
-    }
-}
-
-impl std::error::Error for RegenerateError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        Some(&self.source)
-    }
-}
-
-impl RegenerationStateLease {
-    pub(crate) fn state_mut(&mut self) -> &mut RegenerationState {
-        self.state.as_mut().unwrap()
-    }
-
-    pub(crate) fn commit(mut self) {
-        self.keep_caches = true;
-    }
-}
-
-impl Drop for RegenerationStateLease {
-    fn drop(&mut self) {
-        let mut state = self.state.take().unwrap();
-        if !self.keep_caches && state.caches_dirty {
-            state.glyph_cache = GlyphCache::default();
-            if state.ttf_cache.is_some() {
-                state.ttf_cache = Some(TtfGlyphCache::default());
-            }
-        }
-        state.caches_dirty = false;
-        *self.slot.lock().unwrap() = Some(state);
-    }
-}
-
-/// Result of a successful `generateWebfonts` call. Exposes the generated
-/// font bytes (or `null` for formats that were not requested) and methods to
-/// render the CSS and HTML preview.
-#[cfg_attr(feature = "napi", napi)]
-pub struct GenerateWebfontsResult {
-    pub(crate) cached: OnceLock<Result<CachedTemplateData, String>>,
-    /// Render-cache entries carried across an incremental `regenerate` to seed the rebuilt
-    /// [`CachedTemplateData`], so CSS/HTML that
-    /// doesn't depend on what changed isn't re-rendered. `None` for a normal build.
-    pub(crate) carried_render: Option<CarriedRenderCache>,
-    pub(crate) css_context: Option<Map<String, Value>>,
-    pub(crate) fonts: FontOutputs,
-    pub(crate) html_context: Option<Map<String, Value>>,
-    pub(crate) options: Arc<ResolvedGenerateWebfontsOptions>,
-    pub(crate) regeneration_state: Arc<Mutex<Option<RegenerationState>>>,
-    pub(crate) source_files: Arc<Vec<LoadedSvgFile>>,
-}
-
-// Pure Rust getters (always available)
-impl GenerateWebfontsResult {
-    #[cfg(any(feature = "napi", feature = "bench"))]
-    pub(crate) fn snapshot_for_regeneration(&self, state: RegenerationState) -> Self {
-        Self {
-            cached: OnceLock::new(),
-            carried_render: self.render_cache_source(),
-            css_context: self.css_context.clone(),
-            fonts: self.fonts.clone(),
-            html_context: self.html_context.clone(),
-            options: self.options.clone(),
-            regeneration_state: Arc::new(Mutex::new(Some(state))),
-            source_files: self.source_files.clone(),
-        }
-    }
-
-    pub(crate) fn take_regeneration_state(&self) -> std::io::Result<RegenerationState> {
-        if !self.options.incremental {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::InvalidInput,
-                "regenerate requires the font to be generated with `incremental` enabled.",
-            ));
-        }
-        self.regeneration_state
-            .lock()
-            .unwrap()
-            .take()
-            .ok_or_else(|| {
-                std::io::Error::new(
-                    std::io::ErrorKind::WouldBlock,
-                    "This result is already regenerating or has been replaced.",
-                )
-            })
-    }
-
-    pub(crate) fn take_regeneration_state_lease(&self) -> std::io::Result<RegenerationStateLease> {
-        Ok(RegenerationStateLease {
-            slot: Arc::clone(&self.regeneration_state),
-            state: Some(self.take_regeneration_state()?),
-            keep_caches: false,
-        })
-    }
-
-    #[cfg(feature = "bench")]
-    pub(crate) fn restore_regeneration_state(&self, state: RegenerationState) {
-        *self.regeneration_state.lock().unwrap() = Some(state);
-    }
-
-    pub(crate) fn seed_written_outputs(&self, written_outputs: HashMap<String, [u8; 16]>) {
-        if let Some(state) = self.regeneration_state.lock().unwrap().as_mut() {
-            state.written_outputs = written_outputs;
-        }
-    }
-
-    #[doc(hidden)]
-    #[cfg(feature = "bench")]
-    pub fn regenerate_owned_for_bench(
-        &self,
-        files: &[String],
-        changes: &[(String, GlyphChange)],
-    ) -> std::io::Result<Self> {
-        let state = self.take_regeneration_state()?;
-        let mut replacement = self.snapshot_for_regeneration(state);
-        match replacement.regenerate(files, changes) {
-            Ok(()) => Ok(replacement),
-            Err(error) => {
-                if let Ok(state) = replacement.take_regeneration_state() {
-                    self.restore_regeneration_state(state);
-                }
-                Err(error)
-            }
-        }
-    }
-
-    /// Returns the EOT font bytes, if generated.
-    pub fn eot_bytes(&self) -> Option<&[u8]> {
-        self.fonts.eot_font.as_ref().map(|v| v.as_ref().as_slice())
-    }
-
-    /// Returns the SVG font string, if generated.
-    pub fn svg_string(&self) -> Option<&str> {
-        self.fonts.svg_font.as_ref().map(|v| v.as_ref().as_str())
-    }
-
-    /// Returns the TTF font bytes, if generated.
-    pub fn ttf_bytes(&self) -> Option<&[u8]> {
-        self.fonts.ttf_font.as_ref().map(|v| v.as_ref().as_slice())
-    }
-
-    /// Returns the WOFF font bytes, if generated.
-    pub fn woff_bytes(&self) -> Option<&[u8]> {
-        self.fonts.woff_font.as_ref().map(|v| v.as_ref().as_slice())
-    }
-
-    /// Returns the WOFF2 font bytes, if generated.
-    pub fn woff2_bytes(&self) -> Option<&[u8]> {
-        self.fonts
-            .woff2_font
-            .as_ref()
-            .map(|v| v.as_ref().as_slice())
-    }
-}
-
-// NAPI getters and methods
-#[cfg(feature = "napi")]
-#[napi]
-impl GenerateWebfontsResult {
-    /// EOT font bytes, or `null` if EOT was not in `types`.
-    #[napi(getter)]
-    pub fn eot(&self) -> Option<Uint8Array> {
-        self.fonts
-            .eot_font
-            .as_ref()
-            .map(|v| Uint8Array::from(v.as_ref().clone()))
-    }
-
-    /// SVG font XML string, or `null` if SVG was not in `types`.
-    #[napi(getter)]
-    pub fn svg(&self) -> Option<String> {
-        self.fonts.svg_font.as_ref().map(|v| v.as_ref().clone())
-    }
-
-    /// TTF font bytes, or `null` if TTF was not in `types`.
-    #[napi(getter)]
-    pub fn ttf(&self) -> Option<Uint8Array> {
-        self.fonts
-            .ttf_font
-            .as_ref()
-            .map(|v| Uint8Array::from(v.as_ref().clone()))
-    }
-
-    /// WOFF2 font bytes, or `null` if WOFF2 was not in `types`.
-    #[napi(getter)]
-    pub fn woff2(&self) -> Option<Uint8Array> {
-        self.fonts
-            .woff2_font
-            .as_ref()
-            .map(|v| Uint8Array::from(v.as_ref().clone()))
-    }
-
-    /// WOFF font bytes, or `null` if WOFF was not in `types`.
-    #[napi(getter)]
-    pub fn woff(&self) -> Option<Uint8Array> {
-        self.fonts
-            .woff_font
-            .as_ref()
-            .map(|v| Uint8Array::from(v.as_ref().clone()))
-    }
-
-    /// Render the CSS string for this result. Pass `urls` to override the
-    /// default font URLs in the `@font-face src:` descriptor (only the keys
-    /// you supply are overridden). The result is cached per `urls` value, so
-    /// repeated calls with the same input are cheap.
-    #[napi(ts_args_type = "urls?: Partial<Record<FontType, string>>")]
-    pub fn generate_css(&self, urls: Option<HashMap<String, String>>) -> napi::Result<String> {
-        let urls = urls.map(parse_native_urls).transpose()?;
-        self.generate_css_pure(urls).map_err(to_napi_err)
-    }
-
-    /// Render the HTML preview string for this result. Pass `urls` to
-    /// override font URLs in the embedded stylesheet (only the keys you
-    /// supply are overridden). The result is cached per `urls` value.
-    #[napi(ts_args_type = "urls?: Partial<Record<FontType, string>>")]
-    pub fn generate_html(&self, urls: Option<HashMap<String, String>>) -> napi::Result<String> {
-        let urls = urls.map(parse_native_urls).transpose()?;
-        self.generate_html_pure(urls).map_err(to_napi_err)
-    }
-
-    /// Rebuild the font after a batch of file changes, reusing cached glyph geometry for files
-    /// whose contents are unchanged. Requires the font to have been generated with
-    /// `incremental: true`. `files` is the complete file set after the changes, in the order a
-    /// fresh build would use (e.g. the glob result) — the rebuilt glyphs are ordered to match it,
-    /// so the output bytes are identical to a fresh `generateWebfonts` of that set. `changes`
-    /// describes the affected files: added/changed files are re-read from disk; any file absent
-    /// from `files` is dropped. Omit `changes` to re-read/hash every current file and infer
-    /// added/changed/removed paths from `files`. Every requested format is refreshed in memory,
-    /// and — like `generateWebfonts` — when the result was built with `writeFiles` enabled the
-    /// refreshed fonts are written to disk too, while CSS/HTML companion files are skipped if their
-    /// rendered bytes are unchanged since the last write.
-    #[napi(js_name = "regenerate")]
-    pub fn regenerate_from_js(
-        &mut self,
-        files: Vec<String>,
-        changes: Option<Vec<GlyphChangeEntry>>,
-    ) -> napi::Result<()> {
-        let changes = parse_glyph_changes(changes)?;
-        match changes {
-            Some(changes) => self.regenerate(&files, &changes),
-            None => self.regenerate_all(&files),
-        }
-        .map_err(to_napi_err)
-    }
-
-    /// Rebuild off the Node.js event loop and resolve with a replacement result. The receiver
-    /// remains readable and unchanged while regeneration runs and after failure. Assign the
-    /// resolved result before starting another regeneration. Overlapping calls on the same result
-    /// lineage are rejected, and disk writes remain non-transactional.
-    #[napi(js_name = "regenerateAsync")]
-    pub async fn regenerate_async_from_js(
-        &self,
-        files: Vec<String>,
-        changes: Option<Vec<GlyphChangeEntry>>,
-    ) -> napi::Result<GenerateWebfontsResult> {
-        let changes = parse_glyph_changes(changes)?;
-        let state = self.take_regeneration_state().map_err(to_napi_err)?;
-        let original_state = Arc::clone(&self.regeneration_state);
-        let mut replacement = self.snapshot_for_regeneration(state);
-        tokio::task::spawn_blocking(move || -> std::io::Result<GenerateWebfontsResult> {
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| match changes {
-                Some(changes) => replacement.regenerate(&files, &changes),
-                None => replacement.regenerate_all(&files),
-            }));
-            match result {
-                Ok(Ok(())) => Ok(replacement),
-                Ok(Err(error)) => {
-                    if let Ok(state) = replacement.take_regeneration_state() {
-                        *original_state.lock().unwrap() = Some(state);
-                    }
-                    Err(error)
-                }
-                Err(payload) => {
-                    if let Ok(state) = replacement.take_regeneration_state() {
-                        *original_state.lock().unwrap() = Some(state);
-                    }
-                    std::panic::resume_unwind(payload)
-                }
-            }
-        })
-        .await
-        .map_err(|error| {
-            napi::Error::from_reason(format!("Native webfont regeneration task failed: {error}"))
-        })?
-        .map_err(to_napi_err)
-    }
-}
-
-#[cfg(feature = "napi")]
-fn to_napi_err(error: impl std::fmt::Display) -> napi::Error {
-    napi::Error::new(napi::Status::GenericFailure, error.to_string())
-}
-
-#[cfg(feature = "napi")]
-fn parse_glyph_changes(
-    changes: Option<Vec<GlyphChangeEntry>>,
-) -> napi::Result<Option<Vec<(String, GlyphChange)>>> {
-    changes
-        .map(|changes| {
-            changes
-                .into_iter()
-                .map(|entry| {
-                    let change = match entry.change_type.as_str() {
-                        "added" => GlyphChange::Added { name: entry.name },
-                        "changed" => GlyphChange::Changed { name: entry.name },
-                        "removed" => GlyphChange::Removed,
-                        other => {
-                            return Err(napi::Error::from_reason(format!(
-                                "Unknown changeType '{other}'; expected 'added', 'changed', or 'removed'."
-                            )));
-                        }
-                    };
-                    Ok((entry.path, change))
-                })
-                .collect()
-        })
-        .transpose()
-}
-
-#[cfg(feature = "napi")]
-fn parse_native_urls(urls: HashMap<String, String>) -> napi::Result<HashMap<FontType, String>> {
-    urls.into_iter()
-        .filter_map(|(font_type, url)| {
-            let font_type = match font_type.as_str() {
-                "svg" => Some(FontType::Svg),
-                "ttf" => Some(FontType::Ttf),
-                "eot" => Some(FontType::Eot),
-                "woff" => Some(FontType::Woff),
-                "woff2" => Some(FontType::Woff2),
-                _ => None,
-            }?;
-
-            Some(Ok((font_type, url)))
-        })
-        .collect::<napi::Result<HashMap<FontType, String>>>()
 }
